modify data type
subject table 에 x, y의 data type 수정. (int -> double)

+++ DB/DDL.sql
@@ -15,8 +15,8 @@ CREATE TABLE professor(
 CREATE TABLE subject(
     id varchar(25),
     name varchar(25),
-    x INT,
-    y INT,
+    x double,
+    y double,
     randomnum varchar(25),
     pid varchar(25)
 );
